Comment on a file that might need changing.
I believe this is the file that does the addition for the campaign.  If 
we added on line 225 newIndex = newIndex + increment - opted_out=true.  
If this is the right object to collect that information, we might be 
able to cache the new number without constant updates.  At the moment a 
Texting End User clicks to get more assigned to their batch, it would 
simultaneously get an updated number on contacts in the list that has 
opted out from other campaigns would get subtracted from the total 
number.

Co-Authored-By: Schuyler Duveen <sky@skyb.us>